Move enr version insertion to up a layer to discv5
