+++ src/discv5/test.rs
@@ -1,6 +1,9 @@
 #![cfg(test)]
 
-use crate::{kbucket, Discv5, *};
+use crate::{
+    discv5::{CHECK_VERSION, NAT, TOPICS},
+    kbucket, Discv5, *,
+};
 use enr::{k256, CombinedKey, Enr, EnrBuilder, EnrKey, NodeId};
 use rand_core::{RngCore, SeedableRng};
 use std::{collections::HashMap, net::Ipv4Addr};
@@ -623,4 +626,21 @@ async fn test_bucket_limits() {
 
     // Number of entries should be equal to `bucket_limit`.
     assert_eq!(discv5.kbuckets.read().iter_ref().count(), bucket_limit);
+}
+
+#[test]
+fn test_version_check() {
+    // Create the test values needed
+    let port = 6666;
+    let ip: std::net::IpAddr = "127.0.0.1".parse().unwrap();
+    let key = CombinedKey::generate_secp256k1();
+    let mut enr = crate::enr::EnrBuilder::new("v4")
+        .ip(ip)
+        .udp4(port)
+        .build(&key)
+        .unwrap();
+    let supported_versions = TOPICS | NAT;
+    enr.insert("version", &[supported_versions], &key).unwrap();
+
+    assert!(CHECK_VERSION(&enr, vec!(supported_versions)));
 }
